+++ darts/ad/anomaly_model/forecasting_am.py
@@ -508,7 +508,7 @@ class ForecastingAnomalyModel(AnomalyModel):
             Receiver Operating Characteristic Curve) and "AUC_PR" (Average Precision from scores).
             Default: "AUC_ROC".
         component_wise
-            If True, it will separately plot each component in multivariate series.
+            If True, will separately plot each component in case of multivariate anomaly detection.
         score_kwargs
             parameters for the `score()` method.
         """